+++ packages/ag-charts-enterprise/src/series/ohlc/ohlcSeriesBase.ts
@@ -195,20 +195,10 @@ export abstract class OhlcSeriesBase<
 
     override getSeriesRange(_direction: _ModuleSupport.ChartAxisDirection, visibleRange: [any, any]): [number, number] {
         const { dataModel, processedData } = this;
-        if (!dataModel || !processedData) return [NaN, NaN];
+        const xRange = this.getXRangeInVisibleRange(visibleRange);
+        if (!dataModel || !processedData || !xRange) return [NaN, NaN];
 
-        const xAxis = this.axes[ChartAxisDirection.X]!;
-        const xScale = xAxis.scale;
-        const xValues = dataModel.resolveKeysById(this, `xValue`, processedData);
-
-        const barWidth = xScale.bandwidth ?? 0;
-
-        const [x0, x1] = visibleRangeIndices(xValues.length, visibleRange, (index) => {
-            const x = xScale.convert(xValues[index]);
-            return [x, x + barWidth];
-        });
-
-        return dataModel.getDomainBetweenRange(this, ['highValue', 'lowValue'], [x0, x1], processedData);
+        return dataModel.getDomainBetweenRange(this, ['highValue', 'lowValue'], xRange, processedData);
     }
 
     override createNodeData() {
